Change the ABI for "join" on variants (#199)
This commit changes the ABI for variants such as `union { f32, f64 }`
which in the prior version of the canonical ABI would look like `i32
f64` but with the current canonical ABI draft it would instead look like
`i32 i64`. The previous `unify` function is updated to the new name used
in the canonical ABI explainer, `join`, the contents are modified, and
one form of bitcast was removed as a result.

Additionally I updated the bless'd output for `wasmlink` to ensure
everything has newlines at the end since I was getting some otherwise
odd changes locally. This is hopefully a one-time change for those
baselines.